refactor(ir): move constraint metadata from schedule/ to ir/constraints/
ScheduleConstraint, ScheduleConstraintMetadata and the Layout / Mesh /
StorageConstraint types are authored metadata: the parser produces them from a
where(...) annotation and attaches them to an IR Expr (they are IRMetadata
subclasses, and IRMetadata itself lives in ir/core/metadata.py). They are
consumed by parser, inspection, analysis and target alike, which makes them
IR-layer vocabulary rather than solving-layer vocabulary.

After the move to ir/constraints/ (a peer of ir/types/), analysis/ imports
nothing from schedule/, so the one-way layering (Schedule -> Analysis) holds
unconditionally. The class names are unchanged, to keep a rename out of the
8 referencing files. 361 tests.
