+++ src/tilefoundry/ir/constraints/base.py
